+++ datatest/src/commonMain/kotlin/at/asitplus/test/Data.kt
@@ -24,12 +24,6 @@ object DataTest {
      */
     var defaultDisplayNameMaxLength: Int = -1
 
-    @Deprecated("to be removed in 0.6", replaceWith = ReplaceWith("defaultTestNameMaxLength"), level = DeprecationLevel.ERROR)
-    var maxLength
-        get() = defaultTestNameMaxLength
-        set(value) {
-            defaultTestNameMaxLength = value
-        }
 }